fix: upsert/update data ttl is integer

+++ src/main/java/com/volcengine/vikingdb/runtime/vector/model/request/UpsertDataRequest.java
@@ -23,7 +23,7 @@ public class UpsertDataRequest extends BaseCollectionRequest {
     private List<Map<String, Object>> data;
 
     @JsonProperty("ttl")
-    private Long ttl;
+    private Integer ttl;
 
     @JsonProperty("ignore_unknown_fields")
     private Boolean ignoreUnknownFields;
